Single indentation level for hanging indent for function calls.
This addresses Issue #1, that the double-indentation for
line-continuation in function or method calls causes style warnings with tools
like Pylint and pep8. Even though such indent is still PEP-8 compliant,
the tool authors hold strong opinions that this must be one indent
level. Since convention is in the favor of these tools, I have
introduced this change.